refactor: move Parity to quantum_numbers
